fixed spacing on db reveiw and add dynmaic route on load for single page
