+++ app/components/SingleProductContainer.js
@@ -1,32 +1,28 @@
-import { connect } from 'react-redux';
-import { browserHistory } from 'react-router';
-import React from 'react';
+import Reat from 'react'
+import { connect } from 'react-redux'
+import { Link } from 'react-router'
+import SingleProduct from './SingleProduct'
 
+import {addReviewThunk} from '../reducers/meme'
 
-class SingleProductContainer extends React.Component {
-
-	render(){
-	console.log('selected meme inside of single product container', this.props)
-		return (
-			<div className="create-account-container container"> 
-				<p>{this.props.selectedMeme.name}</p>
-			
-			
-			</div>
-		)
-	}
+const mapDispatchToProps = function(dispatch){
+  return {
+    onSubmit: function(obj) {
+      dispatch(addReviewThunk(obj))
+    }
+  }
 }
 
-const mapStateToProps = (state, ownProps) => {
-	return {
-		memes: state.meme.memes,
-		selectedMeme: state.meme.selectedMeme
-	}
-	
+const mapStateToProps = function(state) {
+  return {
+    reviews: state.meme.reviews,
+    selectedmeme: state.meme.selectedMeme
+  }
 }
 
-const mapDispatchToProps = null;
+
+const ConnectedSingleProducet = connect(mapStateToProps, mapDispatchToProps)(SingleProduct)
+
+export default ConnectedSingleProducet
 
 
-
-export default connect(mapStateToProps, mapDispatchToProps)(SingleProductContainer)
